:lipstick: Remove theme contrast selection
