+++ app/src/commonMain/kotlin/com/crosspaste/ui/theme/ThemeDetector.kt
@@ -15,13 +15,10 @@ interface ThemeDetector {
     )
 
     fun setThemeColor(themeColor: ThemeColor)
-
-    fun setColorContrast(colorContrast: ColorContrast)
 }
 
 data class ThemeState(
     val themeColor: ThemeColor,
-    val colorContrast: ColorContrast,
     val isFollowSystem: Boolean,
     val isUserInDark: Boolean,
     val isSystemInDark: Boolean,
@@ -31,7 +28,6 @@ data class ThemeState(
     companion object {
         fun createThemeState(
             themeColor: ThemeColor,
-            colorContrast: ColorContrast,
             isFollowSystem: Boolean,
             isUserInDark: Boolean,
             isSystemInDark: Boolean,
@@ -39,39 +35,18 @@ data class ThemeState(
             val isCurrentThemeDark = if (isFollowSystem) isSystemInDark else isUserInDark
             val currentColorScheme =
                 if (isCurrentThemeDark) {
-                    getDarkColorSchema(themeColor, colorContrast)
+                    themeColor.darkColorScheme
                 } else {
-                    getLightColorSchema(themeColor, colorContrast)
+                    themeColor.lightColorScheme
                 }
             return ThemeState(
                 themeColor = themeColor,
-                colorContrast = colorContrast,
                 isFollowSystem = isFollowSystem,
                 isUserInDark = isUserInDark,
                 isSystemInDark = isSystemInDark,
                 colorScheme = currentColorScheme,
             )
         }
-
-        private fun getLightColorSchema(
-            themeColor: ThemeColor,
-            contrast: ColorContrast,
-        ): ColorScheme =
-            when (contrast) {
-                ColorContrast.Standard -> themeColor.lightColorScheme
-                ColorContrast.Medium -> themeColor.lightMediumContrastColorScheme
-                ColorContrast.High -> themeColor.lightHighContrastColorScheme
-            }
-
-        private fun getDarkColorSchema(
-            themeColor: ThemeColor,
-            contrast: ColorContrast,
-        ): ColorScheme =
-            when (contrast) {
-                ColorContrast.Standard -> themeColor.darkColorScheme
-                ColorContrast.Medium -> themeColor.darkMediumContrastColorScheme
-                ColorContrast.High -> themeColor.darkHighContrastColorScheme
-            }
     }
 
     val isCurrentThemeDark: Boolean
