+++ app/src/desktopMain/kotlin/com/crosspaste/ui/theme/DesktopThemeDetector.kt
@@ -17,12 +17,10 @@ class DesktopThemeDetector(
 ) : ThemeDetector {
 
     private val initialThemeColor = CrossPasteTheme.getThemeColor(configManager.getCurrentConfig().themeColor)
-    private val initialColorContrast = ColorContrast.valueOf(configManager.getCurrentConfig().colorContrast)
     private val initialFollowSystem = configManager.getCurrentConfig().isFollowSystemTheme
     private val initialIsDarkTheme = configManager.getCurrentConfig().isDarkTheme
 
     private val _themeColor = MutableStateFlow(initialThemeColor)
-    private val _colorContrast = MutableStateFlow(initialColorContrast)
     private val _isFollowSystem = MutableStateFlow(initialFollowSystem)
     private val _isUserInDark = MutableStateFlow(initialIsDarkTheme)
     private val _isSystemInDark = MutableStateFlow(false)
@@ -30,15 +28,13 @@ class DesktopThemeDetector(
     override val themeState: StateFlow<ThemeState> =
         combine(
             _themeColor,
-            _colorContrast,
             _isFollowSystem,
             _isUserInDark,
             _isSystemInDark,
-        ) { color, contrast, follow, user, system ->
+        ) { color, follow, user, system ->
 
             createThemeState(
                 themeColor = color,
-                colorContrast = contrast,
                 isFollowSystem = follow,
                 isUserInDark = user,
                 isSystemInDark = system,
@@ -49,7 +45,6 @@ class DesktopThemeDetector(
             initialValue =
                 createThemeState(
                     themeColor = initialThemeColor,
-                    colorContrast = initialColorContrast,
                     isFollowSystem = initialFollowSystem,
                     isUserInDark = initialIsDarkTheme,
                     isSystemInDark = false,
@@ -76,10 +71,5 @@ class DesktopThemeDetector(
     override fun setThemeColor(themeColor: ThemeColor) {
         _themeColor.value = themeColor
         configManager.updateConfig("themeColor", themeColor.name)
-    }
-
-    override fun setColorContrast(colorContrast: ColorContrast) {
-        _colorContrast.value = colorContrast
-        configManager.updateConfig("colorContrast", colorContrast.name)
     }
 }
